Moved Http2 OutputQueueRecord into separate class.
